% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tile_grids.R
\name{create_tile_catalog}
\alias{create_tile_catalog}
\title{Create a tile catalog data frame}
\usage{
create_tile_catalog(tiles_vect)
}
\arguments{
\item{tiles_vect}{SpatVector with tiles}
}
\value{
data.frame with tile metadata
}
\description{
Create a tile catalog data frame
}
