% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_tile_identification.R
\name{create_tile_templates_from_df}
\alias{create_tile_templates_from_df}
\title{Create tile templates only for identified tiles}
\usage{
create_tile_templates_from_df(tile_df, zones)
}
\arguments{
\item{tile_df}{data.frame from fast_identify_tiles with tile_id, zone_id, col, row}

\item{zones}{UTM zone definitions}
}
\value{
List of SpatRaster templates
}
\description{
Create tile templates only for identified tiles
}
