% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_tile_identification.R
\name{empty_tile_df}
\alias{empty_tile_df}
\title{Empty tile data.frame with the canonical columns}
\usage{
empty_tile_df()
}
\description{
Empty tile data.frame with the canonical columns
}
\keyword{internal}
