% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tile_grids.R
\name{export_tile_catalog}
\alias{export_tile_catalog}
\title{Export tile catalog to CSV}
\usage{
export_tile_catalog(tiles_vect, filename)
}
\arguments{
\item{tiles_vect}{SpatVector with tiles}

\item{filename}{Output CSV filename}
}
\description{
Export tile catalog to CSV
}
