% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tile_grids.R
\name{generate_tile_hierarchy}
\alias{generate_tile_hierarchy}
\title{Generate matching L1 and L2 tile sets}
\usage{
generate_tile_hierarchy(feature_vect, zones)
}
\arguments{
\item{feature_vect}{SpatVector to cover}

\item{zones}{UTM zone definitions}
}
\value{
list with L1 and L2 tile sets
}
\description{
Generate matching L1 and L2 tile sets
}
