% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tile_grids.R
\name{get_aat_regions}
\alias{get_aat_regions}
\title{Define extents for key AAT regions}
\usage{
get_aat_regions()
}
\value{
list of extents in terra ordering c(xmin, xmax, ymin, ymax)
}
\description{
Define extents for key AAT regions
}
