% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctic_grid_system.R
\name{load_grid_spec}
\alias{load_grid_spec}
\title{Load grid specification from RDS file}
\usage{
load_grid_spec(filename = "antarctic_grid_spec.rds")
}
\description{
Load grid specification from RDS file
}
