% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctic_grid_system.R
\name{lonlat_to_utm}
\alias{lonlat_to_utm}
\title{Convert lon/lat coordinates to UTM in a given zone}
\usage{
lonlat_to_utm(lon, lat, zone)
}
\arguments{
\item{lon, lat}{Numeric vectors, WGS84 degrees}

\item{zone}{Zone identifier, e.g. "43S" (see \code{\link[=define_utm_zones]{define_utm_zones()}})}
}
\value{
data.frame with x, y (UTM metres)
}
\description{
Convert lon/lat coordinates to UTM in a given zone
}
