% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctic_grid_system.R
\name{pad_zone_id}
\alias{pad_zone_id}
\title{Zero-pad a zone identifier ("1S" -> "01S"); vectorised}
\usage{
pad_zone_id(zone_id)
}
\description{
Zero-pad a zone identifier ("1S" -> "01S"); vectorised
}
\keyword{internal}
