% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctic_grid_system.R
\name{save_grid_spec}
\alias{save_grid_spec}
\title{Save grid specification to RDS file}
\usage{
save_grid_spec(filename = "antarctic_grid_spec.rds")
}
\description{
Save grid specification to RDS file
}
