% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs.R
\name{utm_extent_to_lonlat}
\alias{utm_extent_to_lonlat}
\title{Lonlat bbox of a UTM extent, densified (cross-zone helper)}
\usage{
utm_extent_to_lonlat(extent, epsg, n = 21)
}
\description{
Lonlat bbox of a UTM extent, densified (cross-zone helper)
}
\keyword{internal}
