% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_tile_identification.R
\name{cells_to_tile_indices}
\alias{cells_to_tile_indices}
\title{Get tile indices from raster cell indices}
\usage{
cells_to_tile_indices(zone_raster, cell_indices)
}
\arguments{
\item{zone_raster}{SpatRaster with tiles as cells}

\item{cell_indices}{Vector of cell indices}
}
\value{
data.frame with cell, col, row indices
}
\description{
Get tile indices from raster cell indices
}
