% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_utm_zone_boundaries.R
\name{generate_utm_zone_boundaries}
\alias{generate_utm_zone_boundaries}
\title{Generate UTM zone boundary lines}
\usage{
generate_utm_zone_boundaries(
  zone_numbers,
  lat_range = c(-85, -40),
  n_points = 100
)
}
\arguments{
\item{zone_numbers}{Vector of UTM zone numbers (default: 42:58 for AAT)}

\item{lat_range}{Vector c(min_lat, max_lat) in degrees (default: c(-85, -40))}

\item{n_points}{Number of points along each line (default: 100)}
}
\value{
SpatVector of lines in EPSG:4326
}
\description{
Generate UTM zone boundary lines
}
