% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_tile_identification.R
\name{get_tile_extents_from_df}
\alias{get_tile_extents_from_df}
\title{Get tile extents for identified tiles}
\usage{
get_tile_extents_from_df(tile_df, zones)
}
\arguments{
\item{tile_df}{data.frame from fast_identify_tiles}

\item{zones}{UTM zone definitions}
}
\value{
data.frame with tile_id and extent columns (xmin, xmax, ymin, ymax)
}
\description{
Get tile extents for identified tiles
}
